(refactor): `_compat.sum` to `_utils.elem_sum`
