+++ scanpy/preprocessing/_utils.py
@@ -7,8 +7,7 @@ import numpy as np
 from scipy import sparse
 from sklearn.random_projection import sample_without_replacement
 
-from .._compat import sum
-from .._utils import AnyRandom, _SupportedArray, elem_mul
+from .._utils import AnyRandom, _SupportedArray, elem_mul, elem_sum
 
 if TYPE_CHECKING:
     import dask.array as da
@@ -16,7 +15,7 @@ if TYPE_CHECKING:
 
 
 def dask_array_mean(X: da.Array, axis):
-    total = sum(X, axis=axis)
+    total = elem_sum(X, axis=axis)
     return total / X.shape[axis]
