+++ scanpy/tests/test_normalization.py
@@ -10,7 +10,6 @@ from scipy import sparse
 from scipy.sparse import csr_matrix
 
 import scanpy as sc
-from scanpy._compat import sum
 from scanpy.testing._helpers import (
     _check_check_values_warnings,
     check_rep_mutation,
@@ -32,13 +31,13 @@ X_frac = np.array([[1, 0, 1], [3, 0, 1], [5, 6, 1]])
 def test_normalize_total(array_type, dtype):
     adata = AnnData(array_type(X_total).astype(dtype))
     sc.pp.normalize_total(adata, key_added="n_counts")
-    assert np.allclose(np.ravel(sum(adata.X, axis=1)), [3.0, 3.0, 3.0])
+    assert np.allclose(np.ravel(sc._utils.elem_sum(adata.X, axis=1)), [3.0, 3.0, 3.0])
     sc.pp.normalize_total(adata, target_sum=1, key_added="n_counts2")
-    assert np.allclose(np.ravel(sum(adata.X, axis=1)), [1.0, 1.0, 1.0])
+    assert np.allclose(np.ravel(sc._utils.elem_sum(adata.X, axis=1)), [1.0, 1.0, 1.0])
 
     adata = AnnData(array_type(X_frac).astype(dtype))
     sc.pp.normalize_total(adata, exclude_highly_expressed=True, max_fraction=0.7)
-    assert np.allclose(np.ravel(sum(adata.X, axis=1))[:2], [2.0, 4.0])
+    assert np.allclose(np.ravel(sc._utils.elem_sum(adata.X, axis=1))[:2], [2.0, 4.0])
 
 
 @pytest.mark.parametrize("array_type", ARRAY_TYPES)
@@ -57,7 +56,9 @@ def test_normalize_total_layers(array_type, dtype):
     adata.layers["layer"] = adata.X.copy()
     with pytest.warns(FutureWarning, match=r".*layers.*deprecated"):
         sc.pp.normalize_total(adata, layers=["layer"])
-    assert np.allclose(sum(adata.layers["layer"], axis=1), [3.0, 3.0, 3.0])
+    assert np.allclose(
+        sc._utils.elem_sum(adata.layers["layer"], axis=1), [3.0, 3.0, 3.0]
+    )
 
 
 @pytest.mark.parametrize("array_type", ARRAY_TYPES)
